# public/swagger.yaml
openapi: 3.0.0
info:
  title: 'Therapeutic Radiesthesia API'
  description: '`API para demonstrar endpoints básicos do sistema`<br><ul><li>Essa documentação está sendo desenvolvida gradualmente, todos os endpoints estão passando por uma revisão.</li><li>Os endpoints que precisam estar com a autorização estão com um cadeado indicando o uso.</li><li>Para gerar o token, utilize o endpoint login.</li><li>Endpoints marcados como acesso ADMIN.</li></ul>'
  contact:
    name: 'Paulo Henrique'
    url: 'https://api.conect.app'
    email: webmaster@multidesk.io
  version: 1.0.0
servers:
  -
    url: 'https://api.multidesk.io'
    description: 'Servidor online'
  -
    url: 'http://localhost:8000'
    description: 'Servidor local'
paths:
...
  /api/v1/login:
    post:
      tags:
        - Autenticação
      summary: Login
      description: 'Retorna um token JWT'
      operationId: 3b889298fc72067506f70fb423b6a004
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: usuario@exemplo.com
                password:
                  type: string
                  format: password
                  example: '123456'
              type: object
      responses:
        '200':
          description: 'Token JWT gerado com sucesso'
          content:
            application/json:
              schema:
                properties:
                  token: { type: string, example: eyJhbGciOiJIUzI1NiIsInR... }
                type: object
        '401':
          description: 'Credenciais inválidas'
        '400':
          description: 'Dados de entrada inválidos'
  /api/v1/magiclink:
    post:
      tags:
        - Autenticação
      summary: 'Autenticação com Link Mágico'
      description: 'Autentica o usuário utilizando um token de link mágico e retorna um token JWT.'
      operationId: cb7d7f1d97cbed904cb95ea42ba6d90a
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - magiclink
              properties:
                magiclink:
                  type: string
                  example: token_de_link_magico_gerado
              type: object
      responses:
        '200':
          description: 'Token JWT gerado com sucesso'
          content:
            application/json:
              schema:
                properties:
                  token: { type: string, example: eyJhbGciOiJIUzI1NiIsInR... }
                type: object
        '400':
          description: 'Token de link mágico não fornecido ou inválido'
        '500':
          description: 'Erro inesperado no servidor'
  /api/v1/logout:
    get:
      tags:
        - Autenticação
      summary: 'Logout do usuário'
      description: 'Realiza logout e invalida o token JWT'
      operationId: f9413ca466b6033b3b76dbc7823c0d85
      responses:
        '200':
          description: 'Logout realizado com sucesso'
        '401':
          description: 'Token inválido ou ausente'
      security:
        -
          bearerAuth: []
  /api/v1/recover:
    put:
      tags:
        - Autenticação
      summary: 'Atualização de senha'
      description: 'Permite que o usuário atualize sua senha usando um token de recuperação.'
      operationId: 1e6906d84277eab763bcc0b24db707ae
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - token
                - password
              properties:
                token:
                  type: string
                  example: token_de_recuperacao_gerado
                password:
                  type: string
                  example: NovaSenhaForte123!
              type: object
      responses:
        '200':
          description: 'Senha atualizada com sucesso'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Password updated successfully' }
                type: object
        '400':
          description: 'Token ou senha não fornecidos'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'Token and password required' }
                type: object
        '500':
          description: 'Erro inesperado no servidor'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'Internal server error' }
                type: object
    post:
      tags:
        - Autenticação
      summary: 'Recuperação de senha'
      description: 'Permite que um usuário solicite a recuperação de senha através do seu email.'
      operationId: cfcb64eeb4b59dfcb0b78b6e80c1feff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
              type: object
      responses:
        '200':
          description: 'Recuperação de senha bem-sucedida'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Recovery successfully completed' }
                  data: { type: object, example: '{"recovery_token": "xyz123abc"}' }
                type: object
        '400':
          description: 'Email não fornecido ou inválido'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'Email is required' }
                type: object
        '404':
          description: 'Usuário não encontrado'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'User not found' }
                type: object
        '500':
          description: 'Erro inesperado no servidor'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'Internal server error' }
                type: object
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...